refactor: move more files around
