--- examples/web/src/declarations/typographic-base.d.ts
@@ -1,1 +1,0 @@
-declare module 'typographic-base'
