+++ packages/api-reference/src/components/DarkModeToggle/DarkModeIconToggle.vue
