+++ packages/api-reference/src/components/DarkModeToggle/DarkModeToggle.vue
