+++ packages/api-reference/src/components/DarkModeToggle/index.ts
@@ -1,0 +1,1 @@
+export { default as DarkModeToggle } from './DarkModeToggle.vue'
