+++ packages/api-reference/src/components/GettingStarted.vue
@@ -3,8 +3,8 @@ import { ScalarButton } from '@scalar/components'
 import { type ThemeId } from '@scalar/themes'
 import { ref, watch } from 'vue'
 
-import coinmarketcap from '../coinmarketcapv3.json'
-import petstore from '../petstorev3.json'
+import coinmarketcap from '../specs/coinmarketcapv3.json'
+import petstore from '../specs/petstorev3.json'
 import { type GettingStartedExamples } from '../types'
 
 const props = defineProps<{
