+++ packages/api-reference/src/specs/coinmarketcapv3.json
