+++ packages/api-reference/src/specs/petstorev3.json
