+++ packages/api-reference/src/standalone.ts
@@ -1,3 +1,7 @@
+/**
+ * This file is the entry point for the CDN version of the API Reference.
+ * It’s responsible for finding the spec and configuration in the HTML, and mounting the Vue.js app.
+ */
 import { createApp } from 'vue'
 
 import { default as ApiReference } from './components/ApiReference.vue'
