+++ packages/api-reference/src/types.ts
@@ -1,12 +1,7 @@
 import { type ThemeId } from '@scalar/themes'
 import type { MetaFlatInput } from '@unhead/schema'
 import type { HarRequest } from 'httpsnippet-lite'
-import {
-  type OpenAPI,
-  type OpenAPIV2,
-  type OpenAPIV3,
-  type OpenAPIV3_1,
-} from 'openapi-types'
+import { type OpenAPIV2, type OpenAPIV3, type OpenAPIV3_1 } from 'openapi-types'
 import { type DeepReadonly, type Slot } from 'vue'
 
 export type ReferenceProps = {
